refactor: fix esling errors
