--- src/app.js
@@ -1,3 +1,0 @@
-/* eslint-disable linebreak-style */
-
-// import Keyboard from './keyboard';
