+++ webpack.config.js
@@ -1,3 +1,4 @@
+/* eslint-disable linebreak-style */
 /* eslint-disable indent */
 /* eslint-disable linebreak-style */
 
@@ -5,17 +6,21 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-    entry: './src/keyboard.js', 
+    entry: './src/keyboard.js',
     module: {
         rules: [
-            { test: /\.css$/, use: [ 'style-loader', 'css-loader' ] },
-            { test: /\.(js)$/, use: 'babel-loader' }
-        ]
+            {
+ test: /\.css$/,
+                use: ['style-loader', 'css-loader'],
+},
+            {
+ test: /\.(js)$/,
+                use: 'babel-loader',
+},
+        ],
     },
-    
     output: {
-      path: path.resolve(__dirname, 'dist'), // Директория, в которой будет
-      // размещаться итоговый бандл, папка dist в корне приложения
+      path: path.resolve(__dirname, 'dist'),
       filename: 'index_bundle.js',
       clean: {
         keep: /\.git/,
@@ -25,4 +30,4 @@ module.exports = {
         new HtmlWebpackPlugin(),
       ],
       mode: 'production',
-}
+};
